+++ packages/iov-keycontrol/src/keyring.ts
@@ -1,9 +1,26 @@
 import { As } from "type-tagger";
 
-import { Wallet, WalletId, WalletImplementationIdString, WalletSerializationString } from "./wallet";
+import { ChainId, PublicIdentity } from "@iov/bcp";
+import { Ed25519Keypair, Slip10RawIndex } from "@iov/crypto";
+
+import {
+  ReadonlyWallet,
+  Wallet,
+  WalletId,
+  WalletImplementationIdString,
+  WalletSerializationString,
+} from "./wallet";
 import { Ed25519HdWallet, Ed25519Wallet, Secp256k1HdWallet } from "./wallets";
 
 export type KeyringSerializationString = string & As<"keyring-serialization">;
+
+/**
+ * Read-only information about one wallet in a keyring
+ */
+export interface WalletInfo {
+  readonly id: WalletId;
+  readonly label: string | undefined;
+}
 
 interface WalletSerialization {
   readonly implementationId: WalletImplementationIdString;
@@ -82,26 +99,64 @@ export class Keyring {
     }
   }
 
-  public add(wallet: Wallet): void {
+  public add(wallet: Wallet): WalletInfo {
     this.wallets.push(wallet);
+    return {
+      id: wallet.id,
+      label: wallet.label.value,
+    };
   }
 
   /**
-   * this returns an array with mutable element references. Thus e.g.
-   * .getWallets().createIdentity() will change the keyring.
+   * Returns an array with immutable references.
    */
-  public getWallets(): ReadonlyArray<Wallet> {
+  public getWallets(): ReadonlyArray<ReadonlyWallet> {
     return this.wallets;
   }
 
   /**
-   * Finds a wallet and returns a mutable references. Thus e.g.
-   * .getWallet(xyz).createIdentity() will change the keyring.
+   * Finds a wallet and returns an immutable references.
    *
    * @returns a wallet if ID is found, undefined otherwise
    */
-  public getWallet(id: WalletId): Wallet | undefined {
+  public getWallet(id: WalletId): ReadonlyWallet | undefined {
     return this.wallets.find(wallet => wallet.id === id);
+  }
+
+  /** Sets the label of the wallet with the given ID in the primary keyring  */
+  public setWalletLabel(walletId: WalletId, label: string | undefined): void {
+    const wallet = this.getMutableWallet(walletId);
+    if (!wallet) {
+      throw new Error(`Wallet of id '${walletId}' does not exist in keyring`);
+    }
+    wallet.setLabel(label);
+  }
+
+  /**
+   * Creates an identitiy in the wallet with the given ID in the primary keyring
+   *
+   * The identity is bound to one chain ID to encourage using different
+   * keypairs on different chains.
+   */
+  public async createIdentity(
+    walletId: WalletId,
+    chainId: ChainId,
+    options: Ed25519Keypair | ReadonlyArray<Slip10RawIndex> | number,
+  ): Promise<PublicIdentity> {
+    const wallet = this.getMutableWallet(walletId);
+    if (!wallet) {
+      throw new Error(`Wallet of id '${walletId}' does not exist in keyring`);
+    }
+    return wallet.createIdentity(chainId, options);
+  }
+
+  /** Assigns a label to one of the identities in the wallet with the given ID in the primary keyring */
+  public setIdentityLabel(walletId: WalletId, identity: PublicIdentity, label: string | undefined): void {
+    const wallet = this.getMutableWallet(walletId);
+    if (!wallet) {
+      throw new Error(`Wallet of id '${walletId}' does not exist in keyring`);
+    }
+    wallet.setIdentityLabel(identity, label);
   }
 
   // serialize will produce a representation that can be writen to disk.
@@ -121,5 +176,15 @@ export class Keyring {
 
   public clone(): Keyring {
     return new Keyring(this.serialize());
+  }
+
+  /**
+   * Finds a wallet and returns a mutable references. Thus e.g.
+   * .getMutableWallet(xyz).createIdentity(...) will change the keyring.
+   *
+   * @returns a wallet if ID is found, undefined otherwise
+   */
+  private getMutableWallet(id: WalletId): Wallet | undefined {
+    return this.wallets.find(wallet => wallet.id === id);
   }
 }