+++ packages/iov-keycontrol/types/keyring.d.ts
@@ -1,6 +1,15 @@
 import { As } from "type-tagger";
-import { Wallet, WalletId, WalletImplementationIdString, WalletSerializationString } from "./wallet";
+import { ChainId, PublicIdentity } from "@iov/bcp";
+import { Ed25519Keypair, Slip10RawIndex } from "@iov/crypto";
+import { ReadonlyWallet, Wallet, WalletId, WalletImplementationIdString, WalletSerializationString } from "./wallet";
 export declare type KeyringSerializationString = string & As<"keyring-serialization">;
+/**
+ * Read-only information about one wallet in a keyring
+ */
+export interface WalletInfo {
+    readonly id: WalletId;
+    readonly label: string | undefined;
+}
 export declare type WalletDeserializer = (data: WalletSerializationString) => Wallet;
 /**
  * A collection of wallets
@@ -11,19 +20,35 @@ export declare class Keyring {
     private static deserializeWallet;
     private readonly wallets;
     constructor(data?: KeyringSerializationString);
-    add(wallet: Wallet): void;
+    add(wallet: Wallet): WalletInfo;
     /**
-     * this returns an array with mutable element references. Thus e.g.
-     * .getWallets().createIdentity() will change the keyring.
+     * Returns an array with immutable references.
      */
-    getWallets(): ReadonlyArray<Wallet>;
+    getWallets(): ReadonlyArray<ReadonlyWallet>;
     /**
-     * Finds a wallet and returns a mutable references. Thus e.g.
-     * .getWallet(xyz).createIdentity() will change the keyring.
+     * Finds a wallet and returns an immutable references.
      *
      * @returns a wallet if ID is found, undefined otherwise
      */
-    getWallet(id: WalletId): Wallet | undefined;
+    getWallet(id: WalletId): ReadonlyWallet | undefined;
+    /** Sets the label of the wallet with the given ID in the primary keyring  */
+    setWalletLabel(walletId: WalletId, label: string | undefined): void;
+    /**
+     * Creates an identitiy in the wallet with the given ID in the primary keyring
+     *
+     * The identity is bound to one chain ID to encourage using different
+     * keypairs on different chains.
+     */
+    createIdentity(walletId: WalletId, chainId: ChainId, options: Ed25519Keypair | ReadonlyArray<Slip10RawIndex> | number): Promise<PublicIdentity>;
+    /** Assigns a label to one of the identities in the wallet with the given ID in the primary keyring */
+    setIdentityLabel(walletId: WalletId, identity: PublicIdentity, label: string | undefined): void;
     serialize(): KeyringSerializationString;
     clone(): Keyring;
+    /**
+     * Finds a wallet and returns a mutable references. Thus e.g.
+     * .getMutableWallet(xyz).createIdentity(...) will change the keyring.
+     *
+     * @returns a wallet if ID is found, undefined otherwise
+     */
+    private getMutableWallet;
 }